move catshadow.Contact sort.Interface into sort.go
