+++ contact.go
@@ -17,7 +17,6 @@ import (
 	"gioui.org/widget/material"
 	"github.com/benc-uk/gofract/pkg/colors"
 	"github.com/benc-uk/gofract/pkg/fractals"
-	"github.com/katzenpost/katzenpost/catshadow"
 	"github.com/katzenpost/hpqc/rand"
 	qrcode "github.com/skip2/go-qrcode"
 	"golang.org/x/exp/shiny/materialdesign/icons"
@@ -26,7 +25,6 @@ import (
 	"io"
 	mrand "math/rand"
 	"runtime"
-	"sort"
 	"strings"
 	"sync"
 )
@@ -354,41 +352,6 @@ func (p *AddContactPage) layoutQr(gtx C) D {
 	t.Pop()
 	return dims
 
-}
-
-type sortedContacts []*catshadow.Contact
-
-func (s sortedContacts) Less(i, j int) bool {
-	// sorts contacts with messages most-recent-first, followed by contacts
-	// without messages alphabetically
-	if s[i].LastMessage == nil && s[j].LastMessage == nil {
-		return s[i].Nickname < s[j].Nickname
-	} else if s[i].LastMessage == nil {
-		return false
-	} else if s[j].LastMessage == nil {
-		return true
-	} else {
-		return s[i].LastMessage.Timestamp.After(s[j].LastMessage.Timestamp)
-	}
-}
-func (s sortedContacts) Swap(i, j int) {
-	s[i], s[j] = s[j], s[i]
-}
-func (s sortedContacts) Len() int {
-	return len(s)
-}
-
-func getSortedContacts(a *App) (contacts sortedContacts) {
-	if a.c == nil {
-		return
-	}
-
-	// returns map[string]*Contact
-	for _, contact := range a.c.GetContacts() {
-		contacts = append(contacts, contact)
-	}
-	sort.Sort(contacts)
-	return
 }
 
 func button(th *material.Theme, button *widget.Clickable, icon *widget.Icon) material.IconButtonStyle {
